+++ Source/AnimusExMachina/Private/AI/AxMAIController.cpp
@@ -35,6 +35,11 @@ void AAxMAIController::PostInitializeComponents()
 {
 	Super::PostInitializeComponents();
 
+	if (StateTreeAI)
+	{
+		StateTreeAI->PrimaryComponentTick.TickInterval = StateTreeTickInterval;
+	}
+
 	if (!PerceptionComp)
 	{
 		return;